Fix HDF5-related undefined behavior
This commit fixes several undefined- and implementation-defined
behaviors in libcbf, identified from an analysis of warnings
emitted by GCC.  The motivation was to address crashes in the
nexus2cbf example program that exhibit a sensitivity to (at
least) compiler optimization level. Specific issues addressed
include

 - left-shifting negative integers
 - pointer arithmetic involving type void *
 - calling a function that has no in-scope declaration
 - string / character-array manipulation with dependencies on
   uninitialized, unset contents of local arrays
 - computing a pointer to before the start of an object
 - accessing an object (via pointer) after the end of the
   object's lifetime

It was the last of those, in function _cbf_nx2cbfDread_scalar_string,
that seems to have been the main source of the specific problem
observed.

Fixes #15